requirements.txt
totalsegmentator
tensorly
fvcore
tqdm
SimpleITK
scipy
torchinfo
git+https://github.com/deepmind/surface-distance.git
